Rework the galileo(1) man page
Remove the introduction and feature section, this is not the goal of the
man page, people looking at this page already installed galileo and know
what it is good for, they are just looking for more informations about
hidden options, or further pointers.

The --config option specified an extra configuration file

The --no-https-only only enable the fallback, https remains the default.